Always define noExitRuntime
Previously we when EXIT_RUNTIME was set we were leaving it as undefined.
This change explicitly initializes it.

The old behavior had causes some odd behavior previously in Fetch.js
where noExitRuntime was not set at the start of a fetch request (as
would be expected).

Also, minor formatting cleanup of module receiving code which I noticed
was pretty hard to read in the final output file.