Remove EM_EXCLUSIVE_CACHE_ACCESS. NFC
This is my second attempt and removing this and I think
this time I've got it right this time.

The cache lock should never normally be needed in recursive
builds so assert if a parent process holds the lock. See
the comment as to why this should never happen (and if it
does its a bug).